fix(tests): relocate dot-assignment fixtures out of the P-4 scan surface (#243)
CI's p4_runtime_source_scan correctly rejected the synthetic-artifact
test fixtures appended to runtime.rs — the scan covers the whole file,
test code included, and fixture arithmetic (K * D, %, /) is value
arithmetic. Tests move to mod.rs (the scan's own home, unscanned) with
code_from_bundle_dot exposed pub(crate); runtime.rs is again free of
any mul/div/mod token. Full lib suite green locally INCLUDING both P-4
scans this time — the earlier filtered run that excluded the scan is
exactly the cheap-check gap the measurement-discipline rule exists for.
